requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==3.1.2
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
pytest==7.3.1
python-docx==0.8.11
python-dotenv==1.0.0
requests==2.31.0
setuptools==80.9.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zope.event==6.1
zope.interface==8.1
